MS-2 created constants file for top songs lists
