+++ app/src/Pages/Music/index.tsx
@@ -1,27 +1,20 @@
 import React from "react";
 import "../../App.css";
-import { useTranslation } from "react-i18next";
 import { TopSongsList } from "./TopSongsList";
 import { Box } from "@mui/material";
 import { makeStyles } from "@mui/styles";
 import createTheme from "@mui/material/styles/createTheme";
+import { TopDnb2023 } from "../../Constants/topSongsLists";
 
 const Music: React.FC = () => {
-  const { t } = useTranslation("spotifyEmbedLinks");
   const styles = useStyles();
-
-  const songList2023 = [
-    t("imanuOfTwoMinds"),
-    t("gyrofieldFemmeFatale"),
-    t("yaanoHadEnough"),
-    t('signalHowWillIKnow')
-  ];
+  const topDnb2023 = TopDnb2023();
 
   return (
 
     <Box className={styles.listContainer}>
       <Box width='40%'>
-      {songList2023.map(song => (
+      {topDnb2023.map(song => (
       <TopSongsList url={song} />
       ))}
       </Box>
